chore(products): remove addToCart fn
